Make detekt and apiDump happy
