+++ instrumentation/view-click/src/test/kotlin/io/opentelemetry/android/instrumentation/view/click/ViewClickInstrumentationTest.kt
@@ -10,18 +10,18 @@ import android.app.Application
 import android.os.SystemClock
 import android.view.MotionEvent
 import android.view.View
-import android.view.ViewConfiguration
 import android.view.ViewGroup
 import android.view.Window
 import android.view.Window.Callback
-import androidx.test.core.view.PointerCoordsBuilder
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import fastForwardDoubleTapTimeout
+import getDoubleTapSequence
+import getSingleTapSequence
 import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
 import io.mockk.mockk
-import io.mockk.mockkClass
 import io.mockk.slot
 import io.mockk.verify
 import io.opentelemetry.android.instrumentation.InstallationContext
@@ -42,13 +42,12 @@ import io.opentelemetry.semconv.incubating.AppIncubatingAttributes.APP_SCREEN_CO
 import io.opentelemetry.semconv.incubating.AppIncubatingAttributes.APP_SCREEN_COORDINATE_Y
 import io.opentelemetry.semconv.incubating.AppIncubatingAttributes.APP_WIDGET_ID
 import io.opentelemetry.semconv.incubating.AppIncubatingAttributes.APP_WIDGET_NAME
+import mockView
 import org.assertj.core.api.Assertions
 import org.junit.Before
 import org.junit.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.runner.RunWith
-import org.robolectric.shadows.ShadowLooper
-import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 @ExtendWith(MockKExtension::class)
@@ -615,19 +614,11 @@ class ViewClickInstrumentationTest {
             window.callback = capture(wrapperCapturingSlot)
         }
 
-        wrapperCapturingSlot.captured.dispatchTouchEvent(
-            doubleTapSequence[0]
-        )
-        wrapperCapturingSlot.captured.dispatchTouchEvent(
-            doubleTapSequence[1]
-        )
+        wrapperCapturingSlot.captured.dispatchTouchEvent(doubleTapSequence[0])
+        wrapperCapturingSlot.captured.dispatchTouchEvent(doubleTapSequence[1])
         fastForwardDoubleTapTimeout()
-        wrapperCapturingSlot.captured.dispatchTouchEvent(
-            doubleTapSequence[2]
-        )
-        wrapperCapturingSlot.captured.dispatchTouchEvent(
-            doubleTapSequence[3]
-        )
+        wrapperCapturingSlot.captured.dispatchTouchEvent(doubleTapSequence[2])
+        wrapperCapturingSlot.captured.dispatchTouchEvent(doubleTapSequence[3])
 
         val events = openTelemetryRule.logRecords
         Assertions.assertThat(events).hasSize(2)
@@ -784,158 +775,5 @@ class ViewClickInstrumentationTest {
                 equalTo(APP_WIDGET_ID, mockView.id.toString()),
                 equalTo(APP_WIDGET_NAME, "10012"),
             )
-    }
-
-    private inline fun <reified T : View> mockView(
-        id: Int,
-        motionEvent: MotionEvent,
-        hitOffset: IntArray = intArrayOf(0, 0),
-        clickable: Boolean = true,
-        visibility: Int = View.VISIBLE,
-        applyOthers: (T) -> Unit = {},
-    ): T {
-        val mockView = mockkClass(T::class)
-        every { mockView.visibility } returns visibility
-        every { mockView.isClickable } returns clickable
-
-        every { mockView.id } returns id
-        val location = IntArray(2)
-
-        location[0] = (motionEvent.x + hitOffset[0]).toInt()
-        location[1] = (motionEvent.y + hitOffset[1]).toInt()
-
-        val arrayCapturingSlot = slot<IntArray>()
-        every { mockView.getLocationInWindow(capture(arrayCapturingSlot)) } answers {
-            arrayCapturingSlot.captured[0] = location[0]
-            arrayCapturingSlot.captured[1] = location[1]
-        }
-
-        every { mockView.x } returns location[0].toFloat()
-        every { mockView.y } returns location[1].toFloat()
-
-        every { mockView.width } returns (location[0] + hitOffset[0])
-        every { mockView.height } returns (location[1] + hitOffset[1])
-        applyOthers.invoke(mockView)
-
-        return mockView
-    }
-
-    private val allowedToolTypes = arrayOf(MotionEvent.TOOL_TYPE_FINGER, MotionEvent.TOOL_TYPE_MOUSE,
-        MotionEvent.TOOL_TYPE_STYLUS, MotionEvent.TOOL_TYPE_ERASER, MotionEvent.TOOL_TYPE_UNKNOWN)
-
-    private val allowedButtonStates = arrayOf(
-        MotionEvent.BUTTON_PRIMARY, MotionEvent.BUTTON_STYLUS_PRIMARY,
-        MotionEvent.BUTTON_SECONDARY, MotionEvent.BUTTON_STYLUS_SECONDARY,
-        MotionEvent.BUTTON_TERTIARY,
-        MotionEvent.BUTTON_BACK,
-        MotionEvent.BUTTON_FORWARD
-    )
-
-    private fun getDoubleTapSequence(x: Float, y: Float, toolType: Int = MotionEvent.TOOL_TYPE_FINGER, buttonState: Int = 0, exceedTimeOut: Boolean = false)
-    : Array<MotionEvent> {
-
-        require(toolType in allowedToolTypes) {
-            "Invalid tool type"
-        }
-
-        if(buttonState != 0) {
-            require(toolType == MotionEvent.TOOL_TYPE_MOUSE || toolType == MotionEvent.TOOL_TYPE_STYLUS) { "Invalid tool type for button state" }
-            require(buttonState in allowedButtonStates) { "Invalid button state" }
-        }
-
-        val initialTime = SystemClock.uptimeMillis()
-
-        val pointerProperties = MotionEvent.PointerProperties()
-        pointerProperties.id = 0
-        pointerProperties.toolType = toolType
-
-        val pointerCoords = PointerCoordsBuilder.newBuilder().setCoords(x, y).build()
-
-        if(exceedTimeOut) {
-            val doubleTapTimeout = ViewConfiguration.getDoubleTapTimeout()
-
-            return arrayOf(
-                MotionEvent.obtain(initialTime, initialTime,
-                    MotionEvent.ACTION_DOWN, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-                MotionEvent.obtain(initialTime, initialTime + 300L,
-                    MotionEvent.ACTION_UP, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-
-                MotionEvent.obtain(
-                    initialTime + 400L + doubleTapTimeout, initialTime + 500L + doubleTapTimeout,
-                    MotionEvent.ACTION_DOWN, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-
-                MotionEvent.obtain(
-                    initialTime + 600L + doubleTapTimeout, initialTime + 700L + doubleTapTimeout,
-                    MotionEvent.ACTION_UP, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0)
-            )
-        } else {
-
-            return arrayOf(
-                MotionEvent.obtain(initialTime, initialTime,
-                    MotionEvent.ACTION_DOWN, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-                MotionEvent.obtain(initialTime, initialTime + 300L,
-                    MotionEvent.ACTION_UP, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-
-                MotionEvent.obtain(
-                    initialTime + 400L, initialTime + 500L,
-                    MotionEvent.ACTION_DOWN, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0),
-
-                MotionEvent.obtain(
-                    initialTime + 600L, initialTime + 700L,
-                    MotionEvent.ACTION_UP, 1, arrayOf(pointerProperties),
-                    arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                    0, 0, 0, 0)
-            )
-        }
-    }
-
-
-    private fun getSingleTapSequence(x: Float, y: Float, toolType: Int = MotionEvent.TOOL_TYPE_FINGER, buttonState: Int = 0)
-    : Array<MotionEvent> {
-        require(toolType in allowedToolTypes) {
-            "Invalid tool type"
-        }
-
-        if(buttonState != 0) {
-            require(toolType == MotionEvent.TOOL_TYPE_MOUSE || toolType == MotionEvent.TOOL_TYPE_STYLUS) { "Invalid tool type for button state" }
-            require(buttonState in allowedButtonStates) { "Invalid button state" }
-        }
-
-        val initialTime = SystemClock.uptimeMillis()
-
-        val pointerProperties = MotionEvent.PointerProperties()
-        pointerProperties.id = 0
-        pointerProperties.toolType = toolType
-
-        val pointerCoords = PointerCoordsBuilder.newBuilder().setCoords(x, y).build()
-        return arrayOf(
-            MotionEvent.obtain(initialTime, initialTime,
-                MotionEvent.ACTION_DOWN, 1, arrayOf(pointerProperties),
-                arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                0, 0, 0, 0),
-
-            MotionEvent.obtain(initialTime, initialTime + 100L,
-                MotionEvent.ACTION_UP, 1, arrayOf(pointerProperties),
-                arrayOf(pointerCoords), 0, buttonState, 1f, 1f,
-                0, 0, 0, 0)
-        )
-    }
-
-    private fun fastForwardDoubleTapTimeout() {
-        ShadowLooper.idleMainLooper(ViewConfiguration.getDoubleTapTimeout().toLong(), TimeUnit.MILLISECONDS)
     }
 }
